LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-controls-bar-styles-expected.txt

Testing the MacOSFullscreenMediaControls computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS style.bottom is "25px"
PASS style.width is "468px"
PASS style.height is "75px"
PASS bounds.left is 166
PASS bounds.top is 500
PASS bounds.width is 468
PASS bounds.height is 75

PASS successfullyParsed is true

TEST COMPLETE
